lib: introduce string matcher function generator
This commit introduces a case-insensitive string matcher function
generator. It generates (readable) functions that can match strings
faster than converting a string to all lowercase/uppercase and then
checking against a set of strings.

Currently this is used for Buffer.isEncoding() to help improve
performance in the slow case where the encoding does not initially
match a lowered case version.